refactor: Separando servidor de sua inicialização #33
